Some refactoring
